First passing test.
Must use angular-mocks--either in the karma files or required in each
spec--or module won't be found. If added to spec webpack needs to
preprocess those spec files.

The "WARNING: Tried to load angular more than once" is related to adding
angular to the karma files and having it required by app.js. When not in
the karma file list the test fails. >> ELABORATE <<